fixed add to cart product update, renamed editProduct
